Wait for completion of startup/shutdown
Change the plugin to close the Netty channel opened for Redis, and wait
until the channel is closed and the executor terminated.

This resolves a problem when using the plugin in multi-module maven
projects, where sometimes an "address already in use" or similar error
occured when using the plugin in two modules, and the second tried to
start a Redis server before the shutdown of the Redis server from the
first module's build was completed.